pyproject.toml
[project]
name = "inchiki-dns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pycryptodome>=3.22.0",
]
